Do not reset the random state when using pre-fill
By disabling the reset of the random state when using pre-fill and in the
first time step, we achieve a (deterministic) random color for all
pre-filled vehicles.
Due to resetting the state within step 0, all pre-filled vehicles did
not have a random color (or rather the same), since the RNG did not pick
any other numbers in between.