Fix depart position for pre-filled vehicles
round was calculating the wrong bounding box values. Also, the position for pre-filled vehicles can really be random. Therefore, we can simply use uniform to get a pure float value for the depart position. In order to correclty check for the min gap, we include it into bounding boxes for the collision check.